requirements.txt
numpy<2.0
matplotlib
scipy
easyocr
torch 
albucore==0.0.9
albumentations==1.4.8
annotated-types==0.7.0
appnope==0.1.4
asttokens==2.4.1
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
docker-pycreds==0.4.0
executing==2.0.1
filelock==3.14.0
fonttools==4.51.0
fsspec==2024.3.1
gitdb==4.0.11
GitPython==3.1.43
idna==3.7
imageio==2.34.1
ipykernel==6.29.4
ipython==8.24.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jupyter_client==8.6.1
jupyter_core==5.7.2
kiwisolver==1.4.5
lazy_loader==0.4
MarkupSafe==2.1.5
matplotlib==3.8.4
matplotlib-inline==0.1.7
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.3
numpy==1.26.4
opencv-python-headless==4.10.0.82
packaging==24.0
parso==0.8.4
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.2.1
prompt-toolkit==3.0.43
protobuf==5.27.1
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==2.7.4
pydantic_core==2.18.4
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
PyYAML==6.0.1
pyzmq==26.0.3
requests==2.32.3
scikit-image==0.23.2
scikit-learn==1.5.0
scipy==1.13.0
sentry-sdk==2.5.1
setproctitle==1.3.3
setuptools==70.0.0
six==1.16.0
smmap==5.0.1
stack-data==0.6.3
sympy==1.12
threadpoolctl==3.5.0
tifffile==2024.5.10
tomli==2.0.1
torch==2.3.0
torchvision==0.18.0
tornado==6.4
tqdm==4.66.4
traitlets==5.14.3
typing_extensions==4.11.0
urllib3==2.2.1
wandb==0.17.1
wcwidth==0.2.13
numpy